pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-audience-agent"
version = "1.0.0"
description = "AI-powered agent for parsing natural language into audience filters"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AI Audience Agent Team"}
]
keywords = ["ai", "nlp", "langchain", "langgraph", "fastapi", "audience", "filters"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "langchain==0.1.6",
    "langchain-openai==0.0.5",
    "langgraph==0.0.20",
    "langsmith==0.0.87",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "httpx==0.26.0",
    "python-dateutil==2.8.2",
    "structlog==24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==24.1.1",
    "ruff==0.1.14",
    "mypy==1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-audience-agent"
Documentation = "https://github.com/yourusername/ai-audience-agent/docs"
Repository = "https://github.com/yourusername/ai-audience-agent"
"Bug Tracker" = "https://github.com/yourusername/ai-audience-agent/issues"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["app/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

